+++ src/test/java/com/tagmycode/plugin/gui/form/SnippetViewTest.java
@@ -15,7 +15,7 @@ public class SnippetViewTest extends AbstractTestBase {
     public void testConstructor() throws Exception {
         Snippet snippet = new ResourceGenerate().aSnippet();
         snippet.setCode("line1\nline2\nline3\nline4\nline5");
-        SnippetView snippetView = new SnippetView(snippet);
+        SnippetView snippetView = new SnippetView().setSnippet(snippet);
         RTextArea snippetEditorPane = snippetView.getSnippetEditorPane().getTextArea();
         assertEquals("line1\n" +
                 "line2\n" +